Resources/requirements.txt
dateparser==1.1.8
numpy==1.21.6
numpy==1.22.4
opencv_contrib_python==4.6.0.66
opencv_python==4.7.0.72
piexif==1.1.3
Pillow==9.3.0
Pillow==9.4.0
Pillow==8.3.2
Pillow==9.5.0
pipreqs==0.4.13
pytesseract==0.3.10
torch==1.13.1+cu116
ultralytics==8.0.53
